refactor: split dir for CLI and dev scripts
