+++ .github/workflows/get_artifact_dir_name.js
@@ -1,7 +1,7 @@
 const fs = require("fs");
 const os = require("os");
 
-const artifactDirName = require("../../scripts/bin_path").dirName;
+const artifactDirName = require("../../cli/bin_path").dirName;
 
 // Pass artifactDirName to subsequent GitHub actions
 fs.appendFileSync(
