+++ cli/bin_path.js
