+++ cli/rewatch
@@ -5,7 +5,7 @@
 const path = require("path");
 const child_process = require("child_process");
 
-const binAbsolutePath = require("./bin_path").absolutePath;
+const { absolutePath: binAbsolutePath } = require("./bin_path");
 const rewatchExe = path.join(binAbsolutePath, "rewatch.exe");
 
 const args = process.argv.slice(2);
