+++ jscomp/build_tests/case2/input.js
@@ -1,6 +1,6 @@
 var p = require("child_process");
 var assert = require("assert");
-var rescript_exe = require("../../../scripts/bin_path").rescript_exe;
+var { rescript_exe } = require("#cli/bin_path.js");
 var o = p.spawnSync(rescript_exe, { encoding: "utf8", cwd: __dirname });
 
 if (
