+++ jscomp/build_tests/custom_namespace/input.js
@@ -1,6 +1,6 @@
 var child_process = require("child_process");
 var assert = require("assert");
-var rescript_exe = require("../../../scripts/bin_path").rescript_exe;
+var { rescript_exe } = require("#cli/bin_path.js");
 
 child_process.execSync(`${rescript_exe} clean && ${rescript_exe} build`, {
   cwd: __dirname,
