+++ jscomp/build_tests/cycle1/input.js
@@ -3,7 +3,7 @@ const cp = require("child_process");
 const assert = require("assert");
 const fs = require("fs");
 const path = require("path");
-var rescript_exe = require("../../../scripts/bin_path").rescript_exe;
+const { rescript_exe } = require("#cli/bin_path.js");
 
 cp.execSync(`${rescript_exe} clean`, { cwd: __dirname });
