+++ jscomp/build_tests/deprecated-package-specs/input.js
@@ -1,6 +1,6 @@
 const child_process = require("child_process");
 const assert = require("assert");
-const rescript_exe = require("../../../scripts/bin_path").rescript_exe;
+const { rescript_exe } = require("#cli/bin_path.js");
 
 const out = child_process.spawnSync(rescript_exe, { encoding: "utf8" });
 assert.match(
