+++ jscomp/build_tests/devonly/input.js
@@ -1,5 +1,5 @@
 //@ts-check
 var cp = require("child_process");
-var rescript_exe = require("../../../scripts/bin_path").rescript_exe;
+var { rescript_exe } = require("#cli/bin_path.js");
 
 cp.execSync(rescript_exe, { cwd: __dirname, encoding: "utf8" });
