+++ jscomp/build_tests/exports/input.js
@@ -1,5 +1,5 @@
 var child_process = require("child_process");
-var rescript_exe = require("../../../scripts/bin_path").rescript_exe;
+var { rescript_exe } = require("#cli/bin_path.js");
 
 child_process.execSync(rescript_exe, {
   cwd: __dirname,
