+++ jscomp/build_tests/in_source/input.js
@@ -2,7 +2,7 @@ var child_process = require("child_process");
 
 var assert = require("assert");
 
-var rescript_exe = require("../../../scripts/bin_path").rescript_exe;
+var { rescript_exe } = require("#cli/bin_path.js");
 
 assert.throws(
   () => {
