+++ jscomp/build_tests/nested/input.js
@@ -2,7 +2,7 @@ var p = require("child_process");
 var assert = require("assert");
 var fs = require("fs");
 var path = require("path");
-var rescript_exe = require("../../../scripts/bin_path").rescript_exe;
+var { rescript_exe } = require("#cli/bin_path.js");
 p.execSync(rescript_exe, { cwd: __dirname });
 
 var content = fs.readFileSync(path.join(__dirname, "src", "demo.js"), "utf8");
