+++ jscomp/build_tests/ns/input.js
@@ -1,4 +1,4 @@
 var child_process = require("child_process");
-var rescript_exe = require("../../../scripts/bin_path").rescript_exe;
+var { rescript_exe } = require("#cli/bin_path.js");
 
 child_process.execSync(rescript_exe, { cwd: __dirname });
