+++ jscomp/build_tests/post-build/input.js
@@ -1,6 +1,6 @@
 var child_process = require("child_process");
 var assert = require("assert");
-var rescript_exe = require("../../../scripts/bin_path").rescript_exe;
+var { rescript_exe } = require("#cli/bin_path.js");
 
 var out = child_process.spawnSync(rescript_exe, { encoding: "utf8" });
