+++ jscomp/build_tests/rerror/input.js
@@ -1,6 +1,6 @@
 var child_process = require("child_process");
 var assert = require("assert");
-var rescript_exe = require("../../../scripts/bin_path").rescript_exe;
+var { rescript_exe } = require("#cli/bin_path.js");
 child_process.spawnSync(`${rescript_exe} clean`, {
   cwd: __dirname,
   encoding: "utf8",
