--- jscomp/build_tests/transitive_pinned_dependency1/package.json
@@ -1,8 +1,0 @@
-{
-  "name": "top",
-  "workspaces": [
-    "a",
-    "b",
-    "c"
-  ]
-}
