+++ jscomp/build_tests/uncurried-always/input.js
@@ -1,6 +1,6 @@
 //@ts-check
 const cp = require("child_process");
-var rescript_exe = require("../../../scripts/bin_path").rescript_exe;
+var { rescript_exe } = require("#cli/bin_path.js");
 
 cp.execSync(`${rescript_exe} clean`, { cwd: __dirname });
 cp.execSync(`${rescript_exe}`, { cwd: __dirname });
