+++ jscomp/build_tests/unicode/input.js
@@ -1,6 +1,6 @@
 //@ts-check
 var child_process = require("child_process");
-var rescript_exe = require("../../../scripts/bin_path").rescript_exe;
+var { rescript_exe } = require("#cli/bin_path.js");
 
 console.log(child_process.execSync(rescript_exe, { encoding: "utf8" }));
