+++ jscomp/build_tests/warn_legacy_config/input.js
@@ -1,6 +1,6 @@
 const { spawnSync } = require("child_process");
 const assert = require("assert");
-const rescript_exe = require("../../../scripts/bin_path").rescript_exe;
+const { rescript_exe } = require("#cli/bin_path.js");
 
 const output = spawnSync(rescript_exe, { encoding: "utf8" });
 assert(
