+++ jscomp/build_tests/weird_names/input.js
@@ -1,7 +1,7 @@
 var cp = require("child_process");
 var assert = require("assert");
 var path = require("path");
-var rescript_exe = require("../../../scripts/bin_path").rescript_exe;
+var { rescript_exe } = require("#cli/bin_path.js");
 
 var out = cp.spawnSync(rescript_exe, { encoding: "utf8" });
