+++ jscomp/build_tests/weird_names_not_found_bug/input.js
@@ -1,6 +1,6 @@
 var cp = require("child_process");
 var assert = require("assert");
-var rescript_exe = require("../../../scripts/bin_path").rescript_exe;
+var { rescript_exe } = require("#cli/bin_path.js");
 
 var out = cp.spawnSync(rescript_exe, { encoding: "utf8" });
 if (out.stderr !== "") {
