+++ jscomp/build_tests/x-y/input.js
@@ -1,4 +1,4 @@
 var p = require("child_process");
-var rescript_exe = require("../../../scripts/bin_path").rescript_exe;
+var { rescript_exe } = require("#cli/bin_path.js");
 
 p.execSync(rescript_exe);
