+++ jsconfig.json
@@ -6,7 +6,9 @@
     "target": "es2018",
     // target need be set up
     // otherwise some method is missing
-    "module": "commonjs"
+    "module": "nodenext",
     // setup to understand commonjs module
+
+    "moduleResolution": "nodenext"
   }
 }
